requirements.txt
pytz==2023.3
gspread==5.10.0
python-telegram-bot==20.4
functions-framework==3.*
pymongo==4.5.0
